test(radio-card): recategorize and clean up browser tests
